--- a/.env.testing.example
+++ b/.env.testing.example
@@ -13,7 +13,7 @@ DB_HOST=127.0.0.1
 DB_PORT=3306
 DB_DATABASE=sekaijin_test
 DB_USERNAME=root
-DB_PASSWORD=root
+DB_PASSWORD=
 
 # Mapbox Configuration (disabled for tests)
 MAPBOX_ACCESS_TOKEN=
@@ -70,8 +70,8 @@ VITE_PUSHER_APP_CLUSTER="${PUSHER_APP_CLUSTER}"
 MAINTENANCE_MODE=false
 
 # Cloudflare Turnstile - Testing bypass
-TURNSTILE_SITE_KEY=test_key
-TURNSTILE_SECRET_KEY=test_secret
+TURNSTILE_SITE_KEY=test_site_key_for_testing_123
+TURNSTILE_SECRET_KEY=test_secret_key_for_testing_123
 TURNSTILE_BYPASS_LOCAL=true
 
 # Email Verification - Skip in testing