--- a/app/Http/Controllers/AuthController.php
+++ b/app/Http/Controllers/AuthController.php
@@ -3,15 +3,15 @@
 namespace App\Http\Controllers;
 
 use App\Mail\WelcomeEmail;
-use App\Models\User;
 use App\Models\Country;
+use App\Models\User;
+use App\Services\EmailBlacklistService;
+use App\Services\TurnstileService;
 use Illuminate\Auth\Events\Verified;
 use Illuminate\Foundation\Auth\EmailVerificationRequest;
-use App\Services\TurnstileService;
-use App\Services\EmailBlacklistService;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Mail;
@@ -21,12 +21,13 @@ class AuthController extends Controller
 {
     protected TurnstileService $turnstileService;
     protected EmailBlacklistService $blacklistService;
-    
+
     public function __construct(TurnstileService $turnstileService, EmailBlacklistService $blacklistService)
     {
         $this->turnstileService = $turnstileService;
         $this->blacklistService = $blacklistService;
     }
+
     public function showRegister()
     {
         return view('auth.register');
@@ -44,7 +45,7 @@ class AuthController extends Controller
         if ($turnstileCheck !== null) {
             return $turnstileCheck;
         }
-        
+
         // Check email blacklist
         if ($this->blacklistService->isBlacklisted($request->input('email'))) {
             return response()->json([
@@ -52,7 +53,7 @@ class AuthController extends Controller
                 'errors' => ['email' => ['Cette adresse email n\'est pas autorisée. Veuillez utiliser une adresse email valide.']],
             ], 422);
         }
-        
+
         // Validation pour l'étape 1 : création de compte
         $validator = Validator::make($request->all(), [
             'name' => [
@@ -105,13 +106,14 @@ class AuthController extends Controller
         }
 
         // Validate country exists if provided (additional server-side check)
-        if ($request->interest_country && !Country::where('name_fr', $request->interest_country)->exists()) {
+        if ($request->interest_country && ! Country::where('name_fr', $request->interest_country)->exists()) {
             if (request()->ajax() || request()->wantsJson()) {
                 return response()->json([
                     'success' => false,
                     'errors' => ['interest_country' => ['Le pays d\'intérêt sélectionné n\'existe pas.']],
                 ], 422);
             }
+
             return back()->withErrors(['interest_country' => 'Le pays d\'intérêt sélectionné n\'existe pas.'])->withInput();
         }
 
@@ -127,7 +129,7 @@ class AuthController extends Controller
         Auth::login($user);
 
         // Envoyer l'email de vérification (sauf si désactivé en configuration)
-        if (!config('services.email_verification.skip', false)) {
+        if (! config('services.email_verification.skip', false)) {
             try {
                 $user->sendEmailVerificationNotification();
                 \Log::info('Email verification sent successfully', [
@@ -295,7 +297,7 @@ class AuthController extends Controller
         if ($turnstileCheck !== null) {
             return $turnstileCheck;
         }
-        
+
         $credentials = $request->validate([
             'email' => 'required|email',
             'password' => 'required',
@@ -323,25 +325,25 @@ class AuthController extends Controller
 
         return redirect('/')->with('success', 'Vous avez été déconnecté avec succès.');
     }
-    
+
     /**
-     * Verify Turnstile protection for a request
+     * Verify Turnstile protection for a request.
      */
     private function verifyTurnstileForRequest(Request $request, string $action): Response|JsonResponse|null
     {
-        if (!$this->shouldVerifyTurnstile($request)) {
+        if (! $this->shouldVerifyTurnstile($request)) {
             return null;
         }
-        
-        if (!$this->turnstileService->verify($request->input('cf-turnstile-response'), $action)) {
+
+        if (! $this->turnstileService->verify($request->input('cf-turnstile-response'), $action)) {
             \Log::warning('Turnstile verification failed', [
                 'action' => $action,
                 'ip' => $this->anonymizeIp($request->ip()),
-                'token_provided' => !empty($request->input('cf-turnstile-response')),
+                'token_provided' => ! empty($request->input('cf-turnstile-response')),
             ]);
-            
+
             $errorMessage = 'Vérification de sécurité échouée. Veuillez réessayer.';
-            
+
             // Check if request is AJAX
             if ($request->ajax() || $request->wantsJson()) {
                 return response()->json([
@@ -349,23 +351,23 @@ class AuthController extends Controller
                     'errors' => ['turnstile' => [$errorMessage]],
                 ], 422);
             }
-            
+
             return back()->withErrors(['turnstile' => $errorMessage])->withInput();
         }
-        
+
         return null;
     }
-    
+
     /**
-     * Determine if Turnstile verification should be performed
+     * Determine if Turnstile verification should be performed.
      */
     private function shouldVerifyTurnstile(Request $request): bool
     {
-        return $this->turnstileService->isConfigured() && !app()->environment(['local', 'testing']);
+        return $this->turnstileService->isConfigured() && ! app()->environment(['local', 'testing']);
     }
-    
+
     /**
-     * Anonymize IP address for GDPR compliance
+     * Anonymize IP address for GDPR compliance.
      */
     private function anonymizeIp(string $ip): string
     {
@@ -373,18 +375,20 @@ class AuthController extends Controller
         if (filter_var($ip, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4)) {
             $parts = explode('.', $ip);
             $parts[3] = '0';
+
             return implode('.', $parts);
         }
-        
+
         // For IPv6: zero out the last 64 bits
         if (filter_var($ip, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
             $parts = explode(':', $ip);
             for ($i = 4; $i < count($parts); $i++) {
                 $parts[$i] = '0';
             }
+
             return implode(':', $parts);
         }
-        
+
         return 'unknown';
     }
 
@@ -768,21 +772,21 @@ class AuthController extends Controller
 
         try {
             $request->user()->sendEmailVerificationNotification();
-            
+
             \Log::info('Email verification sent successfully', [
                 'user_id' => $request->user()->id,
                 'email' => $request->user()->email,
             ]);
-            
+
             return back()->with('message', 'Email de vérification renvoyé ! Vérifiez votre boîte de réception et vos spams.');
-            
+
         } catch (\Exception $e) {
             \Log::error('Failed to send email verification', [
                 'user_id' => $request->user()->id,
                 'email' => $request->user()->email,
                 'error' => $e->getMessage(),
             ]);
-            
+
             return back()->withErrors(['email' => 'Erreur lors de l\'envoi de l\'email. Veuillez réessayer dans quelques minutes ou contacter le support.']);
         }
     }