--- a/app/Http/Controllers/ProfileController.php
+++ b/app/Http/Controllers/ProfileController.php
@@ -19,14 +19,14 @@ class ProfileController extends Controller
     public function update(Request $request)
     {
         $user = Auth::user();
-        
+
         // Vérifier l'email pour certaines actions critiques (sauf si désactivé en configuration)
-        if (!config('services.email_verification.skip', false) && !$user->hasVerifiedEmail()) {
+        if (! config('services.email_verification.skip', false) && ! $user->hasVerifiedEmail()) {
             // Vérifier directement les champs critiques sans dépendre du client
             $hasAvatarUpload = $request->hasFile('avatar');
             $wantsPublicProfile = $request->boolean('is_public_profile');
-            $hasRestrictedChanges = $hasAvatarUpload || ($wantsPublicProfile && !$user->is_public_profile);
-            
+            $hasRestrictedChanges = $hasAvatarUpload || ($wantsPublicProfile && ! $user->is_public_profile);
+
             if ($hasRestrictedChanges) {
                 return back()->withErrors(['email_verification' => 'Vous devez vérifier votre email avant d\'uploader un avatar ou de rendre votre profil public.'])->withInput();
             }
@@ -564,16 +564,16 @@ class ProfileController extends Controller
 
         if (in_array('is_public_profile', $changedFields)) {
             $isPublicProfile = $request->boolean('is_public_profile', false);
-            
+
             // Vérifier que l'email est vérifié avant de permettre un profil public
-            if ($isPublicProfile && !$user->hasVerifiedEmail()) {
+            if ($isPublicProfile && ! $user->hasVerifiedEmail()) {
                 return response()->json([
                     'success' => false,
                     'errors' => ['is_public_profile' => ['Vous devez vérifier votre email avant de rendre votre profil public.']],
-                    'message' => 'Email non vérifié'
+                    'message' => 'Email non vérifié',
                 ], 422);
             }
-            
+
             $updateData['is_public_profile'] = $isPublicProfile;
         }