… .github/workflows/backup/ci.yml
… .github/workflows/backup/pull-request.yml
… .github/workflows/ci.yml
… app/Models/User.php
… app/Notifications/VerifyEmailNotification.php
… app/Providers/TurnstileServiceProvider.php
… app/Services/EmailBlacklistService.php
… config/app.php
… database/factories/UserFactory.php
… database/migrations/2025_07_21_134315_add_email_verification_indexes_to_users_table.php
… database/seeders/DatabaseSeeder.php
… resources/views/auth/login.blade.php
… resources/views/auth/register.blade.php
… resources/views/contact.blade.php
… resources/views/layout.blade.php
… tests/Feature/ApiEndpointsTest.php
… tests/Feature/AuthenticationTest.php
… tests/Feature/ProfileSimpleTest.php
… tests/Feature/ProfileTest.php
… tests/Feature/RolePermissionTest.php
… tests/Traits/AuthenticatesUsers.php